# feat(filesystem): implement caching to reduce CPU usage (ENG-3408)